+++ service/src/main/java/com/wilma/service/positions/PositionService.java
@@ -2,8 +2,14 @@ package com.wilma.service.positions;
 
 import com.wilma.entity.docs.UserDocument;
 import com.wilma.entity.dto.ApplicationDTO;
+import com.wilma.entity.dto.JobDTO;
+import com.wilma.entity.dto.PlacementDTO;
+import com.wilma.entity.positions.Job;
+import com.wilma.entity.positions.Placement;
 import com.wilma.entity.positions.Position;
 import com.wilma.entity.positions.PositionApplication;
+import com.wilma.repository.JobRepository;
+import com.wilma.repository.PlacementRepository;
 import com.wilma.repository.PositionApplicationRepository;
 import com.wilma.repository.PositionRepository;
 import com.wilma.service.CrudOpsImpl;
@@ -11,9 +17,11 @@ import com.wilma.service.UserService;
 import com.wilma.service.docs.DocumentService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
+import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -27,6 +35,52 @@ public class PositionService extends CrudOpsImpl<Position, Integer, PositionRepo
     private PositionApplicationRepository applicationRepository;
     @Autowired
     private UserService userService;
+
+    @Autowired
+    PositionRepository positionRepository;
+
+    @Autowired
+    JobRepository jobRepository;
+
+    @Autowired
+    PlacementRepository placementRepository;
+
+    public Job addJobFromDTO(JobDTO jobDTO) {
+
+        var currentUser = userService.getCurrentUser();
+        var job = new Job(null, jobDTO.getPartner(), jobDTO.getStartDate(), jobDTO.getEndDate(), jobDTO.getPeriod(), jobDTO.getLocation(), jobDTO.getDescription(), false, false, jobDTO.getPayRate(), jobDTO.getPayType(), jobDTO.getPayFrequency());
+
+        return jobRepository.save(job);
+    }
+
+    public Job updateJobFromDTO(JobDTO jobDTO) {
+        var currentUser = userService.getCurrentUser();
+        var job = new Job(jobDTO.getId(), jobDTO.getPartner(), jobDTO.getStartDate(), jobDTO.getEndDate(), jobDTO.getPeriod(), jobDTO.getLocation(), jobDTO.getDescription(), jobDTO.isFilled(), jobDTO.isApproved(), jobDTO.getPayRate(), jobDTO.getPayType(), jobDTO.getPayFrequency());
+
+        return jobRepository.save(job);
+    }
+
+    public List<Job> getJobs(){
+        return jobRepository.findAll();
+    }
+
+    public Placement addPlacementFromDTO(PlacementDTO placementDTO) {
+        var currentUser = userService.getCurrentUser();
+        var placement = new Placement(null, placementDTO.getPartner(), placementDTO.getStartDate(), placementDTO.getEndDate(), placementDTO.getPeriod(), placementDTO.getLocation(), placementDTO.getDescription(), false, false, false);
+
+        return placementRepository.save(placement);
+    }
+
+    public List<Placement> getPlacements(){
+        return placementRepository.findAll();
+    }
+
+    public Placement updatePlacementFromDTO(PlacementDTO placementDTO) {
+        var currentUser = userService.getCurrentUser();
+        var placement = new Placement(placementDTO.getId(), placementDTO.getPartner(), placementDTO.getStartDate(), placementDTO.getEndDate(), placementDTO.getPeriod(), placementDTO.getLocation(), placementDTO.getDescription(), placementDTO.isFilled(), placementDTO.isApproved(), placementDTO.isCompleted());
+
+        return placementRepository.save(placement);
+    }
 
     /**
      * Submit an application for an available position
